+++ Goodreads.Tests/Clients/FollowersClientTests.cs
@@ -7,7 +7,7 @@ namespace Goodreads.Tests
     public class FollowersClientTests
     {
         private readonly IFollowersClient FollowersClient;
-        private readonly int UserId;
+        private readonly long UserId;
 
         public FollowersClientTests()
         {
