+++ Goodreads/Clients/Implementations/FollowersClient.cs
@@ -26,7 +26,7 @@ namespace Goodreads.Clients
         /// </summary>
         /// <param name="userId">The Goodreads Id for the desired user.</param>
         /// <returns>A Goodreads user following model.</returns>
-        Task<UserFollowingSummary> IFollowersClient.Follow(int userId)
+        Task<UserFollowingSummary> IFollowersClient.Follow(long userId)
         {
             var endpoint = string.Format(CultureInfo.InvariantCulture, "user/{0}/followers", userId);
             return Connection.ExecuteRequest<UserFollowingSummary>(endpoint, null, null, "user-following", Method.POST);
@@ -37,7 +37,7 @@ namespace Goodreads.Clients
         /// </summary>
         /// <param name="userId">The Goodreads Id for the desired user.</param>
         /// <returns>True if the unfollow succeeded, false otherwise.</returns>
-        async Task<bool> IFollowersClient.Unfollow(int userId)
+        async Task<bool> IFollowersClient.Unfollow(long userId)
         {
             var endpoint = string.Format(CultureInfo.InvariantCulture, "user/{0}/followers/stop_following", userId);
             var response = await Connection.ExecuteRaw(endpoint, null, Method.DELETE).ConfigureAwait(false);
