+++ Goodreads/Clients/Interfaces/IAuthorsClient.cs
@@ -13,13 +13,13 @@ namespace Goodreads.Clients
         /// </summary>
         /// <param name="authorId">The Goodreads Id for the desired author.</param>
         /// <returns>An async task returning the desired author information.</returns>
-        Task<Author> GetByAuthorId(int authorId);
+        Task<Author> GetByAuthorId(long authorId);
 
         /// <summary>
         /// Searches Goodreads for the given name and returns an author id if found, null otherwise.
         /// </summary>
         /// <param name="authorName">The author name to search for.</param>
         /// <returns>A Goodreads author id if found, null otherwise.</returns>
-        Task<int?> GetAuthorIdByName(string authorName);
+        Task<long?> GetAuthorIdByName(string authorName);
     }
 }
