+++ Goodreads/Clients/Interfaces/IAuthorsFollowingClient.cs
@@ -13,20 +13,20 @@ namespace Goodreads.Clients
         /// </summary>
         /// <param name="authorId">The Goodreads Id for the desired author.</param>
         /// <returns>A Goodreads author following model.</returns>
-        Task<AuthorFollowing> Follow(int authorId);
+        Task<AuthorFollowing> Follow(long authorId);
 
         /// <summary>
         /// Unfollow an author.
         /// </summary>
         /// <param name="authorFollowingId">The Goodreads Id for the desired author.</param>
         /// <returns>True if the unfollow succeeded, false otherwise.</returns>
-        Task<bool> Unfollow(int authorFollowingId);
+        Task<bool> Unfollow(long authorFollowingId);
 
         /// <summary>
         /// Show author following information.
         /// </summary>
         /// <param name="authorFollowingId">The Goodreads Id for the desired author.</param>
         /// <returns>A Goodreads author following model.</returns>
-        Task<AuthorFollowing> Show(int authorFollowingId);
+        Task<AuthorFollowing> Show(long authorFollowingId);
     }
 }
